chore(notary): handle taxes
