+++ notary/audit/src/lib.rs
@@ -20,6 +20,10 @@ use ulx_notary_primitives::{
 };
 
 pub mod error;
+#[cfg(test)]
+mod test_balanceset;
+#[cfg(test)]
+mod test_notebook;
 
 #[derive(Debug, Clone, PartialEq, TypeInfo, Encode, Decode, Serialize, Deserialize, Snafu)]
 pub enum AccountHistoryLookupError {
@@ -340,15 +344,48 @@ pub struct BalanceChangesetState {
 	/// How much in channel funds was claimed by each account id
 	pub claimed_channel_deposits_per_account: BTreeMap<AccountId, u128>,
 	/// How much tax was sent per account
-	pub tax_per_account: BTreeMap<AccountId, u128>,
+	pub tax_sent_per_account: BTreeMap<AccountId, u128>,
 	/// How much was deposited per account
 	pub claimed_deposits_per_account: BTreeMap<AccountId, u128>,
 
-	restricted_balance: BTreeMap<BTreeSet<(AccountId, AccountType)>, i128>,
+	unclaimed_restricted_balance: BTreeMap<BTreeSet<(AccountId, AccountType)>, i128>,
 	unclaimed_channel_balances: BTreeMap<BTreeSet<AccountId>, i128>,
 }
 
+fn round_up(value: u128, percentage: u128) -> u128 {
+	let numerator = value * percentage;
+
+	let round = if numerator % 100 == 0 { 0 } else { 1 };
+
+	numerator.saturating_div(100) + round
+}
+
 impl BalanceChangesetState {
+	pub fn verify_taxes(&self) -> anyhow::Result<(), VerifyError> {
+		let mut tax_owed_per_account = BTreeMap::new();
+		for (account_id, amount) in self.claimed_deposits_per_account.iter() {
+			let amount = *amount;
+			let tax = if amount < 1000 { round_up(amount, 20) } else { 200 };
+			*tax_owed_per_account.entry(account_id).or_insert(0) += tax;
+		}
+		for (account_id, amount) in self.claimed_channel_deposits_per_account.iter() {
+			let tax = round_up(*amount, 20);
+			*tax_owed_per_account.entry(account_id).or_insert(0) += tax;
+		}
+		for (account_id, tax) in tax_owed_per_account {
+			let tax_sent = self.tax_sent_per_account.get(&account_id).unwrap_or(&0);
+			ensure!(
+				*tax_sent >= tax,
+				VerifyError::InsufficientTaxIncluded {
+					account_id: account_id.clone(),
+					tax_sent: *tax_sent,
+					tax_owed: tax,
+				}
+			);
+		}
+		Ok(())
+	}
+
 	fn verify_change_number(
 		&mut self,
 		change: &BalanceChange,
@@ -386,7 +423,7 @@ impl BalanceChangesetState {
 			for rec in recipients {
 				set.insert((rec.clone(), account_type.clone()));
 			}
-			self.restricted_balance.insert(set, milligons as i128);
+			self.unclaimed_restricted_balance.insert(set, milligons as i128);
 		}
 	}
 
@@ -396,9 +433,7 @@ impl BalanceChangesetState {
 		claimer: &AccountId,
 	) -> anyhow::Result<(), VerifyError> {
 		self.sent_tax += milligons;
-		if let Some(starting) = self.tax_per_account.insert(claimer.clone(), milligons) {
-			self.tax_per_account.get_mut(claimer).map(|a| *a += starting);
-		}
+		*self.tax_sent_per_account.entry(claimer.clone()).or_insert(0) += milligons;
 
 		Ok(())
 	}
@@ -412,16 +447,12 @@ impl BalanceChangesetState {
 		if account_type == &AccountType::Tax {
 			self.claimed_tax += milligons;
 		} else {
-			if let Some(starting) =
-				self.claimed_deposits_per_account.insert(claimer.clone(), milligons)
-			{
-				self.claimed_deposits_per_account.get_mut(claimer).map(|a| *a += starting);
-			}
+			*self.claimed_deposits_per_account.entry(claimer.clone()).or_insert(0) += milligons;
 			self.claimed_deposits += milligons;
 		}
 
 		let key = (claimer.clone(), account_type.clone());
-		self.restricted_balance.retain(|accounts, amount| {
+		self.unclaimed_restricted_balance.retain(|accounts, amount| {
 			if accounts.contains(&key) {
 				let restricted_change = amount.saturating_sub(milligons as i128);
 				if restricted_change > 0 {
@@ -442,13 +473,7 @@ impl BalanceChangesetState {
 		claimer: &AccountId,
 	) -> anyhow::Result<(), VerifyError> {
 		self.claimed_deposits += milligons;
-		if let Some(starting) =
-			self.claimed_channel_deposits_per_account.insert(claimer.clone(), milligons)
-		{
-			self.claimed_channel_deposits_per_account
-				.get_mut(claimer)
-				.map(|a| *a += starting);
-		}
+		*self.claimed_channel_deposits_per_account.entry(claimer.clone()).or_insert(0) += milligons;
 
 		self.unclaimed_channel_balances.retain(|accounts, amount| {
 			if accounts.contains(claimer) {
@@ -511,10 +536,12 @@ impl BalanceChangesetState {
 	}
 }
 
-/// This function verifies the proposed balance changes prior to accessing storage or verifying
+/// This function verifies the proposed balance changes PRIOR to accessing storage or verifying
 /// proofs
 /// 1. Confirm each proposed balance change adds up properly
 /// 2. Confirm the changes net out to 0 (no funds are left outside an account)
+///
+/// Does NOT: lookup anything in storage, verify signatures, or confirm the merkle proofs
 pub fn verify_balance_changeset_allocation(
 	changes: &Vec<BalanceChange>,
 	notebook_number: Option<NotebookNumber>,
@@ -638,7 +665,7 @@ pub fn verify_balance_changeset_allocation(
 		}
 
 		ensure!(
-			balance as u128 == change.balance,
+			balance == change.balance as i128,
 			VerifyError::BalanceChangeMismatch {
 				change_index,
 				provided_balance: change.balance,
@@ -663,1232 +690,7 @@ pub fn verify_balance_changeset_allocation(
 		}
 	);
 	// this works by removing all restricted balances as the approved users draw from them
-	ensure!(state.restricted_balance.is_empty(), VerifyError::InvalidNoteRecipients);
+	ensure!(state.unclaimed_restricted_balance.is_empty(), VerifyError::InvalidNoteRecipients);
 	ensure!(state.unclaimed_channel_balances.is_empty(), VerifyError::InvalidChannelClaimers);
 	Ok(state)
-}
-
-#[cfg(test)]
-mod tests {
-	use std::collections::{BTreeMap, BTreeSet};
-
-	use binary_merkle_tree::{merkle_proof, merkle_root};
-	use chrono::Utc;
-	use codec::Encode;
-	use frame_support::{assert_err, assert_ok, parameter_types};
-	use sp_core::{
-		bounded::BoundedVec, bounded_vec, crypto::AccountId32, sr25519::Signature, Blake2Hasher,
-		H256,
-	};
-	use sp_keyring::{
-		AccountKeyring,
-		Ed25519Keyring::{Dave, Ferdie},
-		Sr25519Keyring::{Alice, Bob},
-	};
-	use sp_runtime::MultiSignature;
-
-	use ulx_notary_primitives::{
-		balance_change::{AccountOrigin, BalanceChange, BalanceProof},
-		note::{AccountType, Note, NoteType},
-		BalanceTip, ChainTransfer, MerkleProof, NewAccountOrigin, Notebook, NotebookHeader,
-		NotebookNumber,
-	};
-
-	use crate::{
-		verify_balance_changeset_allocation, verify_changeset_signatures,
-		verify_previous_balance_proof, AccountHistoryLookupError, NotebookHistoryLookup,
-		VerifyError,
-	};
-
-	use super::notebook_verify;
-
-	fn empty_proof(balance: u128) -> Option<BalanceProof> {
-		Some(BalanceProof {
-			notary_id: 1,
-			notebook_number: 1,
-			balance,
-			notebook_proof: Default::default(),
-			account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-		})
-	}
-
-	fn empty_signature() -> MultiSignature {
-		Signature([0u8; 64]).into()
-	}
-
-	fn test_balance_change(
-		account: AccountKeyring,
-		change_number: u32,
-		balance: u128,
-		prev_balance: u128,
-		milligons: u128,
-		note_type: NoteType,
-	) -> BalanceChange {
-		BalanceChange {
-			account_id: account.to_account_id(),
-			account_type: AccountType::Deposit,
-			change_number,
-			balance,
-			previous_balance_proof: if change_number == 1 {
-				None
-			} else {
-				empty_proof(prev_balance)
-			},
-			channel_hold_note: None,
-			notes: Default::default(),
-			signature: empty_signature(),
-		}
-		.push_note(milligons, note_type)
-		.sign(account.pair())
-		.clone()
-	}
-
-	#[test]
-	fn test_balance_change_allocation_errs_non_zero() {
-		let balance_change = test_balance_change(Alice, 1, 100, 0, 100, NoteType::Claim);
-
-		assert_err!(
-			verify_balance_changeset_allocation(&vec![balance_change], None),
-			VerifyError::BalanceChangeNotNetZero { sent: 0, claimed: 100 }
-		);
-	}
-
-	#[test]
-	fn must_supply_zero_balance_on_first_nonce() {
-		let balance_change = vec![BalanceChange {
-			account_id: Bob.to_account_id(),
-			account_type: AccountType::Deposit,
-			change_number: 2,
-			balance: 100,
-			previous_balance_proof: None,
-			channel_hold_note: None,
-			notes: Default::default(),
-			signature: empty_signature(),
-		}];
-
-		assert_err!(
-			verify_balance_changeset_allocation(&balance_change, None),
-			VerifyError::MissingBalanceProof
-		);
-	}
-
-	#[test]
-	fn test_balance_change_allocation_must_be_zero() {
-		let balance_change = vec![
-			test_balance_change(
-				Bob,
-				2,
-				0,
-				100,
-				100,
-				NoteType::Send { to: Some(bounded_vec!(Alice.to_account_id())) },
-			),
-			test_balance_change(Alice, 1, 100, 0, 100, NoteType::Claim),
-		];
-
-		assert_ok!(verify_balance_changeset_allocation(&balance_change, None));
-	}
-
-	#[test]
-	fn test_notes_must_add_up() {
-		let mut balance_change = vec![
-			BalanceChange {
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 2,
-				balance: 0,
-				previous_balance_proof: empty_proof(250),
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(250, NoteType::Send { to: None })],
-				signature: empty_signature(),
-			},
-			BalanceChange {
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 1,
-				balance: 100,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(100, NoteType::Claim)],
-				signature: empty_signature(),
-			},
-			BalanceChange {
-				account_id: Dave.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 1,
-				balance: 100, // WRONG BALANCE - should be 150
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(150, NoteType::Claim)],
-				signature: empty_signature(),
-			},
-		];
-		assert_err!(
-			verify_balance_changeset_allocation(&balance_change, None),
-			VerifyError::BalanceChangeMismatch {
-				change_index: 2,
-				provided_balance: 100,
-				calculated_balance: 150
-			}
-		);
-
-		balance_change[2].balance = 150;
-		assert_ok!(verify_balance_changeset_allocation(&balance_change, None));
-	}
-
-	#[test]
-	fn test_recipients() {
-		let mut balance_change = vec![
-			BalanceChange {
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 2,
-
-				balance: 0,
-				previous_balance_proof: empty_proof(250),
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(
-					250,
-					NoteType::Send {
-						to: Some(bounded_vec!(Alice.to_account_id(), Ferdie.to_account_id()))
-					}
-				)],
-				signature: empty_signature(),
-			},
-			BalanceChange {
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 1,
-				balance: 200,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(200, NoteType::Claim)],
-				signature: empty_signature(),
-			},
-			BalanceChange {
-				account_id: Dave.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 1,
-				balance: 50,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(50, NoteType::Claim)],
-				signature: empty_signature(),
-			},
-		];
-		assert_err!(
-			verify_balance_changeset_allocation(&balance_change, None),
-			VerifyError::InvalidNoteRecipients
-		);
-
-		balance_change[1].balance = 250;
-		balance_change[1].notes[0].milligons = 250;
-		balance_change.pop();
-		assert_ok!(verify_balance_changeset_allocation(&balance_change, None));
-	}
-
-	#[test]
-	fn test_sending_to_localchain() {
-		let balance_change = vec![BalanceChange {
-			// We look for an transfer to localchain using this id
-			account_id: Bob.to_account_id(),
-			account_type: AccountType::Deposit,
-			change_number: 1,
-			balance: 250,
-			previous_balance_proof: None,
-			channel_hold_note: None,
-			notes: bounded_vec![Note {
-				milligons: 250,
-				note_type: NoteType::ClaimFromMainchain { account_nonce: 1 },
-			}],
-			signature: empty_signature(),
-		}];
-
-		assert_ok!(verify_balance_changeset_allocation(&balance_change, None),);
-	}
-
-	#[test]
-	fn test_sending_to_mainchain() {
-		// This probably never happens - but in this scenario, funds are sent to a localchain to
-		// transfer to a different mainchain account
-		let balance_change = vec![
-			BalanceChange {
-				// We look for an transfer to localchain using this id
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 2,
-				balance: 100,
-				previous_balance_proof: empty_proof(50),
-				channel_hold_note: None,
-				notes: bounded_vec![
-					Note::create(250, NoteType::ClaimFromMainchain { account_nonce: 15 }),
-					Note::create(200, NoteType::Send { to: None }),
-				],
-				signature: empty_signature(),
-			},
-			BalanceChange {
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 1,
-				balance: 50,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![
-					Note::create(200, NoteType::Claim),
-					Note::create(150, NoteType::SendToMainchain),
-				],
-				signature: empty_signature(),
-			},
-		];
-
-		assert_ok!(verify_balance_changeset_allocation(&balance_change, None));
-	}
-
-	#[test]
-	fn test_can_lock_with_a_channel_note() -> anyhow::Result<()> {
-		let channel_note =
-			Note::create(250, NoteType::ChannelHold { recipient: Alice.to_account_id() });
-		let balance_change = BalanceChange {
-			account_id: Bob.to_account_id(),
-			account_type: AccountType::Deposit,
-			change_number: 2,
-			balance: 250,
-			previous_balance_proof: empty_proof(250),
-			channel_hold_note: None,
-			notes: bounded_vec![channel_note.clone()],
-			signature: empty_signature(),
-		};
-		assert_ok!(verify_balance_changeset_allocation(&vec![balance_change], Some(1)),);
-
-		assert_err!(
-			verify_balance_changeset_allocation(
-				&vec![BalanceChange {
-					account_id: Bob.to_account_id(),
-					account_type: AccountType::Deposit,
-					change_number: 3,
-					balance: 250,
-					previous_balance_proof: empty_proof(250),
-					channel_hold_note: Some(channel_note.clone()),
-					notes: bounded_vec![Note::create(250, NoteType::Send { to: None })],
-					signature: empty_signature(),
-				}],
-				Some(2)
-			),
-			VerifyError::AccountLocked
-		);
-
-		let mut channel_settle = BalanceChange {
-			account_id: Bob.to_account_id(),
-			account_type: AccountType::Deposit,
-			change_number: 3,
-			balance: 200,
-			previous_balance_proof: empty_proof(250),
-			channel_hold_note: Some(channel_note.clone()),
-			notes: bounded_vec![Note::create(50, NoteType::ChannelSettle)],
-			signature: empty_signature(),
-		};
-
-		assert_err!(
-			verify_balance_changeset_allocation(&vec![channel_settle.clone()], Some(2)),
-			VerifyError::ChannelHoldNotReadyForClaim
-		);
-
-		// try to clear out change
-		channel_settle.balance = 250;
-		channel_settle.notes[0].milligons = 0;
-
-		// it won't let you claim your own note back
-		assert_err!(
-			verify_balance_changeset_allocation(&vec![channel_settle.clone()], Some(61)),
-			VerifyError::InvalidChannelClaimers
-		);
-
-		// it WILL let you claim back your own note if it's past the grace period
-		assert_ok!(verify_balance_changeset_allocation(&vec![channel_settle.clone()], Some(71)),);
-
-		let changes = vec![
-			BalanceChange {
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 3,
-				balance: 200,
-				previous_balance_proof: empty_proof(250),
-				channel_hold_note: Some(channel_note.clone()),
-				notes: bounded_vec![Note::create(50, NoteType::ChannelSettle)],
-				signature: empty_signature(),
-			},
-			BalanceChange {
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				change_number: 1,
-				balance: 50,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(50, NoteType::ChannelClaim)],
-				signature: empty_signature(),
-			},
-		];
-
-		assert_eq!(
-			verify_balance_changeset_allocation(&changes, None)?.needs_channel_settle_followup,
-			true
-		);
-		// a valid claim is also acceptable
-		assert_ok!(verify_balance_changeset_allocation(&changes, Some(61)),);
-
-		Ok(())
-	}
-
-	#[test]
-	fn test_note_signatures() {
-		let mut balance_change = vec![BalanceChange {
-			// We look for an transfer to localchain using this id
-			account_id: Bob.to_account_id(),
-			account_type: AccountType::Deposit,
-			change_number: 1,
-			balance: 250,
-			previous_balance_proof: None,
-			channel_hold_note: None,
-			notes: bounded_vec![Note::create(
-				250,
-				NoteType::ClaimFromMainchain { account_nonce: 1 }
-			),],
-			signature: empty_signature(),
-		}];
-
-		assert_err!(
-			verify_changeset_signatures(&balance_change),
-			VerifyError::InvalidBalanceChangeSignature { change_index: 0 }
-		);
-
-		balance_change[0].sign(Bob.pair());
-		assert_ok!(verify_changeset_signatures(&balance_change));
-	}
-	#[test]
-	fn test_note_claim_signatures() {
-		let mut balance_change = BalanceChange {
-			account_id: Bob.to_account_id(),
-			account_type: AccountType::Deposit,
-			change_number: 5,
-			previous_balance_proof: empty_proof(250),
-			balance: 250,
-			notes: bounded_vec![],
-			channel_hold_note: None,
-			signature: empty_signature(),
-		};
-		balance_change.push_note(250, NoteType::ChannelHold { recipient: Alice.to_account_id() });
-		balance_change.sign(Bob.pair());
-
-		assert_ok!(verify_changeset_signatures(&vec![balance_change.clone()]));
-
-		let mut balance_change2 = BalanceChange {
-			account_id: Bob.to_account_id(),
-			account_type: AccountType::Deposit,
-			change_number: 6,
-			previous_balance_proof: empty_proof(250),
-			balance: 200,
-			notes: bounded_vec![],
-			channel_hold_note: Some(balance_change.notes[0].clone()),
-			signature: empty_signature(),
-		};
-		balance_change2.push_note(50, NoteType::ChannelSettle);
-		balance_change2.sign(Bob.pair());
-		assert_ok!(verify_changeset_signatures(&vec![balance_change2.clone()]));
-
-		let mut channel_note = balance_change.notes[0].clone();
-
-		balance_change2.sign(Bob.pair());
-		balance_change2.channel_hold_note = Some(Note::create(100, NoteType::Tax));
-		assert_err!(
-			verify_changeset_signatures(&vec![balance_change2.clone()]),
-			VerifyError::InvalidChannelHoldNote
-		);
-
-		channel_note.milligons = 102;
-		balance_change2.channel_hold_note = Some(channel_note.clone());
-		assert_err!(
-			verify_changeset_signatures(&vec![balance_change2.clone()]),
-			VerifyError::InvalidBalanceChangeSignature { change_index: 0 }
-		);
-	}
-
-	#[test]
-	fn test_tax_must_be_claimed_on_tax_account() {
-		let set = vec![
-			BalanceChange {
-				balance: 20_000,
-				change_number: 2,
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: empty_proof(21_000),
-				channel_hold_note: None,
-				notes: Default::default(),
-				signature: empty_signature(),
-			}
-			.push_note(1000, NoteType::Send { to: None })
-			.sign(Bob.pair())
-			.clone(),
-			BalanceChange {
-				balance: 800,
-				change_number: 1,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: None,
-				notes: Default::default(),
-				signature: empty_signature(),
-				channel_hold_note: None,
-			}
-			.push_note(1000, NoteType::Claim)
-			.push_note(200, NoteType::Tax)
-			.sign(Alice.pair())
-			.clone(),
-		];
-
-		assert_err!(
-			verify_balance_changeset_allocation(&set, Some(1)),
-			VerifyError::TaxBalanceChangeNotNetZero { sent: 200, claimed: 0 }
-		);
-
-		let mut claim_tax_on_deposit = set.clone();
-		claim_tax_on_deposit.push(
-			BalanceChange {
-				balance: 200,
-				change_number: 1,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: None,
-				notes: Default::default(),
-				signature: empty_signature(),
-				channel_hold_note: None,
-			}
-			.push_note(200, NoteType::Claim)
-			.clone(),
-		);
-		assert_err!(
-			verify_balance_changeset_allocation(&claim_tax_on_deposit, Some(1)),
-			VerifyError::BalanceChangeNotNetZero { sent: 1000, claimed: 1200 }
-		);
-
-		let mut claim_tax_on_deposit = set.clone();
-		claim_tax_on_deposit.push(
-			BalanceChange {
-				balance: 200,
-				change_number: 1,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Tax,
-				previous_balance_proof: None,
-				notes: Default::default(),
-				signature: empty_signature(),
-				channel_hold_note: None,
-			}
-			.push_note(200, NoteType::Claim)
-			.clone(),
-		);
-		assert_ok!(verify_balance_changeset_allocation(&claim_tax_on_deposit, Some(1)),);
-	}
-
-	struct TestLookup;
-
-	parameter_types! {
-		pub static NotebookRoots: BTreeMap<u32, H256> = BTreeMap::new();
-		pub static LastChangedNotebook: BTreeMap<AccountOrigin, u32> = BTreeMap::new();
-		pub static ValidLocalchainTransfers: BTreeSet<(AccountId32, u32)> = BTreeSet::new();
-	}
-	impl NotebookHistoryLookup for TestLookup {
-		fn get_account_changes_root(
-			_notary_id: u32,
-			notebook_number: NotebookNumber,
-		) -> Result<H256, AccountHistoryLookupError> {
-			NotebookRoots::get()
-				.get(&notebook_number)
-				.ok_or(AccountHistoryLookupError::RootNotFound)
-				.cloned()
-		}
-		fn get_last_changed_notebook(
-			_notary_id: u32,
-			account_origin: AccountOrigin,
-		) -> Result<u32, AccountHistoryLookupError> {
-			LastChangedNotebook::get()
-				.get(&account_origin)
-				.cloned()
-				.ok_or(AccountHistoryLookupError::LastChangeNotFound)
-		}
-		fn is_valid_transfer_to_localchain(
-			_notary_id: u32,
-			account_id: &AccountId32,
-			nonce: u32,
-		) -> Result<bool, AccountHistoryLookupError> {
-			ValidLocalchainTransfers::get()
-				.get(&(account_id.clone(), nonce))
-				.cloned()
-				.ok_or(AccountHistoryLookupError::InvalidTransferToLocalchain)
-				.map(|_| true)
-		}
-	}
-
-	#[test]
-	fn test_verify_previous_balance() {
-		let mut final_balances = BTreeMap::<(AccountId32, AccountType), BalanceTip>::new();
-		let account_id = Alice.to_account_id();
-		let account_type = AccountType::Deposit;
-		let key = (account_id.clone(), account_type.clone());
-
-		let mut change = BalanceChange {
-			account_id,
-			account_type,
-			change_number: 500,
-			balance: 0,
-			previous_balance_proof: None,
-			channel_hold_note: None,
-			notes: bounded_vec![],
-			signature: empty_signature(),
-		};
-		let leaves = vec![
-			BalanceTip {
-				account_id: Dave.to_account_id(),
-				account_type: AccountType::Deposit,
-				balance: 20,
-				change_number: 3,
-				account_origin: AccountOrigin { notebook_number: 5, account_uid: 2 },
-				channel_hold_note: None,
-			}
-			.encode(),
-			BalanceTip {
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Deposit,
-				balance: 100,
-				change_number: 1,
-				account_origin: AccountOrigin { notebook_number: 6, account_uid: 1 },
-				channel_hold_note: None,
-			}
-			.encode(),
-			BalanceTip {
-				account_id: change.account_id.clone(),
-				account_type: change.account_type.clone(),
-				balance: 100,
-				change_number: change.change_number - 1,
-				account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-				channel_hold_note: None,
-			}
-			.encode(),
-		];
-		let merkle_root = merkle_root::<Blake2Hasher, _>(&leaves);
-		NotebookRoots::mutate(|a| {
-			a.insert(7, H256::from_slice([&[0u8], &merkle_root[0..31]].concat().as_ref()))
-		});
-		let origin = AccountOrigin { notebook_number: 1, account_uid: 1 };
-		LastChangedNotebook::mutate(|c| c.insert(origin.clone(), 10));
-
-		let proof = merkle_proof::<Blake2Hasher, _, _>(leaves, 2);
-		change.previous_balance_proof = Some(BalanceProof {
-			notary_id: 1,
-			notebook_number: 7,
-			notebook_proof: Some(MerkleProof {
-				proof: BoundedVec::truncate_from(proof.proof),
-				leaf_index: proof.leaf_index as u32,
-				number_of_leaves: proof.number_of_leaves as u32,
-			}),
-			account_origin: origin.clone(),
-			balance: 100,
-		});
-
-		assert_err!(
-			verify_previous_balance_proof::<TestLookup>(
-				&change.previous_balance_proof.clone().unwrap(),
-				7,
-				&mut final_balances,
-				&change,
-				&key,
-			),
-			VerifyError::InvalidPreviousBalanceChangeNotebook
-		);
-
-		LastChangedNotebook::mutate(|c| c.insert(origin, 7));
-		assert_err!(
-			verify_previous_balance_proof::<TestLookup>(
-				&change.previous_balance_proof.clone().unwrap(),
-				7,
-				&mut final_balances,
-				&change,
-				&key,
-			),
-			VerifyError::InvalidPreviousBalanceProof
-		);
-
-		NotebookRoots::mutate(|a| a.insert(7, merkle_root));
-		assert_ok!(verify_previous_balance_proof::<TestLookup>(
-			&change.previous_balance_proof.clone().unwrap(),
-			7,
-			&mut final_balances,
-			&change,
-			&key,
-		));
-	}
-
-	#[test]
-	fn test_verify_notebook() {
-		let note = Note::create(1000, NoteType::ClaimFromMainchain { account_nonce: 1 });
-
-		let alice_balance_changeset = vec![BalanceChange {
-			balance: 1000,
-			change_number: 1,
-			account_id: Alice.to_account_id(),
-			account_type: AccountType::Deposit,
-			previous_balance_proof: None,
-			channel_hold_note: None,
-			notes: bounded_vec![note],
-			signature: empty_signature(),
-		}
-		.sign(Alice.pair())
-		.clone()];
-		let notebook_header1 = NotebookHeader {
-			version: 1,
-			notary_id: 1,
-			notebook_number: 1,
-			finalized_block_number: 100,
-			pinned_to_block_number: 0,
-			start_time: Utc::now().timestamp_millis() as u64 - 60_000,
-			changed_accounts_root: merkle_root::<Blake2Hasher, _>(vec![BalanceTip {
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				balance: 1000,
-				change_number: 1,
-				account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-				channel_hold_note: None,
-			}
-			.encode()]),
-			chain_transfers: bounded_vec![ChainTransfer::ToLocalchain {
-				account_id: Alice.to_account_id(),
-				account_nonce: 1,
-			}],
-			changed_account_origins: bounded_vec![AccountOrigin {
-				notebook_number: 1,
-				account_uid: 1
-			}],
-			end_time: Utc::now().timestamp_millis() as u64,
-		};
-
-		ValidLocalchainTransfers::mutate(|a| a.insert((Alice.to_account_id(), 1)));
-		let hash = notebook_header1.hash();
-
-		let notebook1 = Notebook {
-			header: notebook_header1.clone(),
-			balance_changes: bounded_vec![BoundedVec::truncate_from(
-				alice_balance_changeset.clone()
-			)],
-			new_account_origins: bounded_vec![NewAccountOrigin::new(
-				Alice.to_account_id(),
-				AccountType::Deposit,
-				1
-			)],
-		};
-
-		assert_ok!(notebook_verify::<TestLookup>(&hash, &notebook1));
-
-		let mut bad_hash = hash.clone();
-		bad_hash.0[0] = 1;
-		assert_err!(
-			notebook_verify::<TestLookup>(&bad_hash, &notebook1),
-			VerifyError::InvalidNotebookHash
-		);
-
-		let mut bad_notebook1 = notebook1.clone();
-		let _ = bad_notebook1.header.chain_transfers.try_insert(
-			0,
-			ChainTransfer::ToLocalchain { account_id: Bob.to_account_id(), account_nonce: 2 },
-		);
-		assert_err!(
-			notebook_verify::<TestLookup>(&hash, &bad_notebook1),
-			VerifyError::InvalidChainTransfersList
-		);
-
-		let mut bad_notebook = notebook1.clone();
-		bad_notebook.header.changed_accounts_root.0[0] = 1;
-		assert_err!(
-			notebook_verify::<TestLookup>(&hash, &bad_notebook),
-			VerifyError::InvalidBalanceChangeRoot
-		);
-	}
-
-	#[test]
-	fn test_disallows_double_claim() {
-		let note1 = Note::create(1000, NoteType::ClaimFromMainchain { account_nonce: 1 });
-		let note2 = Note::create(1000, NoteType::ClaimFromMainchain { account_nonce: 1 });
-
-		let alice_balance_changeset = vec![BalanceChange {
-			balance: 2000,
-			change_number: 1,
-			account_id: Alice.to_account_id(),
-			account_type: AccountType::Deposit,
-			previous_balance_proof: None,
-			channel_hold_note: None,
-			notes: bounded_vec![note1, note2],
-			signature: empty_signature(),
-		}
-		.sign(Alice.pair())
-		.clone()];
-		let notebook_header1 = NotebookHeader {
-			version: 1,
-			notary_id: 1,
-			notebook_number: 1,
-			finalized_block_number: 100,
-			pinned_to_block_number: 0,
-			start_time: Utc::now().timestamp_millis() as u64 - 60_000,
-			changed_accounts_root: merkle_root::<Blake2Hasher, _>(vec![BalanceTip {
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				balance: 2000,
-				change_number: 1,
-				account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-				channel_hold_note: None,
-			}
-			.encode()]),
-			chain_transfers: bounded_vec![ChainTransfer::ToLocalchain {
-				account_id: Alice.to_account_id(),
-				account_nonce: 1,
-			}],
-			changed_account_origins: bounded_vec![AccountOrigin {
-				notebook_number: 1,
-				account_uid: 1
-			}],
-			end_time: Utc::now().timestamp_millis() as u64,
-		};
-
-		ValidLocalchainTransfers::mutate(|a| a.insert((Alice.to_account_id(), 1)));
-		let notebook1 = Notebook {
-			header: notebook_header1.clone(),
-			balance_changes: bounded_vec![BoundedVec::truncate_from(
-				alice_balance_changeset.clone()
-			)],
-			new_account_origins: bounded_vec![NewAccountOrigin::new(
-				Alice.to_account_id(),
-				AccountType::Deposit,
-				1
-			)],
-		};
-		let hash = notebook_header1.hash();
-
-		assert_err!(
-			notebook_verify::<TestLookup>(&hash, &notebook1),
-			VerifyError::DuplicateChainTransfer
-		);
-	}
-
-	#[test]
-	fn test_multiple_changesets_in_a_notebook() {
-		let alice_balance_changeset = vec![
-			BalanceChange {
-				balance: 0,
-				change_number: 1,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![
-					Note::create(1000, NoteType::ClaimFromMainchain { account_nonce: 1 }),
-					Note::create(1000, NoteType::Send { to: None }),
-				],
-				signature: empty_signature(),
-			}
-			.sign(Alice.pair())
-			.clone(),
-			BalanceChange {
-				balance: 800,
-				change_number: 1,
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![
-					Note::create(1000, NoteType::Claim),
-					Note::create(200, NoteType::Tax),
-				],
-				signature: empty_signature(),
-			}
-			.sign(Bob.pair())
-			.clone(),
-			BalanceChange {
-				balance: 200,
-				change_number: 1,
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Tax,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(200, NoteType::Claim),],
-				signature: empty_signature(),
-			}
-			.sign(Bob.pair())
-			.clone(),
-		];
-
-		ValidLocalchainTransfers::mutate(|a| a.insert((Alice.to_account_id(), 1)));
-		// NOTE: this is in sorted order by account_id, account_type
-		let mut balance_tips = BTreeMap::from([
-			(
-				(Alice.to_account_id(), AccountType::Deposit),
-				BalanceTip {
-					account_id: Alice.to_account_id(),
-					account_type: AccountType::Deposit,
-					balance: 0,
-					change_number: 1,
-					account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-					channel_hold_note: None,
-				},
-			),
-			(
-				(Bob.to_account_id(), AccountType::Deposit),
-				BalanceTip {
-					account_id: Bob.to_account_id(),
-					account_type: AccountType::Deposit,
-					balance: 800,
-					change_number: 1,
-					account_origin: AccountOrigin { notebook_number: 1, account_uid: 2 },
-					channel_hold_note: None,
-				},
-			),
-			(
-				(Bob.to_account_id(), AccountType::Tax),
-				BalanceTip {
-					account_id: Bob.to_account_id(),
-					account_type: AccountType::Tax,
-					balance: 200,
-					change_number: 1,
-					account_origin: AccountOrigin { notebook_number: 1, account_uid: 3 },
-					channel_hold_note: None,
-				},
-			),
-		]);
-
-		let mut notebook = Notebook {
-			header: NotebookHeader {
-				version: 1,
-				notary_id: 1,
-				notebook_number: 1,
-				finalized_block_number: 100,
-				pinned_to_block_number: 0,
-				start_time: Utc::now().timestamp_millis() as u64 - 60_000,
-				changed_accounts_root: merkle_root::<Blake2Hasher, _>(
-					balance_tips.iter().map(|(_, v)| v.encode()).collect::<Vec<_>>(),
-				),
-				chain_transfers: bounded_vec![ChainTransfer::ToLocalchain {
-					account_id: Alice.to_account_id(),
-					account_nonce: 1,
-				}],
-				changed_account_origins: bounded_vec![
-					AccountOrigin { notebook_number: 1, account_uid: 1 },
-					AccountOrigin { notebook_number: 1, account_uid: 2 },
-					AccountOrigin { notebook_number: 1, account_uid: 3 }
-				],
-				end_time: Utc::now().timestamp_millis() as u64,
-			},
-			balance_changes: bounded_vec![BoundedVec::truncate_from(alice_balance_changeset),],
-			new_account_origins: bounded_vec![
-				NewAccountOrigin::new(Alice.to_account_id(), AccountType::Deposit, 1),
-				NewAccountOrigin::new(Bob.to_account_id(), AccountType::Deposit, 2),
-				NewAccountOrigin::new(Bob.to_account_id(), AccountType::Tax, 3)
-			],
-		};
-
-		assert_ok!(notebook_verify::<TestLookup>(&notebook.header.hash(), &notebook),);
-
-		let changeset2 = vec![
-			BalanceChange {
-				balance: 0,
-				change_number: 2,
-				account_id: Bob.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(800, NoteType::Send { to: None }),],
-				signature: empty_signature(),
-			}
-			.sign(Bob.pair())
-			.clone(),
-			BalanceChange {
-				balance: 600,
-				change_number: 2,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![
-					Note::create(800, NoteType::Claim),
-					Note::create(200, NoteType::Tax),
-				],
-				signature: empty_signature(),
-			}
-			.sign(Alice.pair())
-			.clone(),
-			BalanceChange {
-				balance: 200,
-				change_number: 1,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Tax,
-				previous_balance_proof: None,
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(200, NoteType::Claim),],
-				signature: empty_signature(),
-			}
-			.sign(Alice.pair())
-			.clone(),
-		];
-		notebook.header.changed_accounts_root = merkle_root::<Blake2Hasher, _>(
-			balance_tips.iter().map(|(_, v)| v.encode()).collect::<Vec<_>>(),
-		);
-		notebook
-			.balance_changes
-			.try_push(BoundedVec::truncate_from(changeset2))
-			.expect("should insert");
-		if let Some(tip) = balance_tips.get_mut(&(Bob.to_account_id(), AccountType::Deposit)) {
-			tip.change_number = 2;
-			tip.balance = 0;
-		}
-		if let Some(tip) = balance_tips.get_mut(&(Alice.to_account_id(), AccountType::Deposit)) {
-			tip.change_number = 2;
-			tip.balance = 600;
-		}
-		balance_tips.insert(
-			(Alice.to_account_id(), AccountType::Tax),
-			BalanceTip {
-				balance: 200,
-				change_number: 1,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Tax,
-				account_origin: AccountOrigin { notebook_number: 1, account_uid: 4 },
-				channel_hold_note: None,
-			},
-		);
-		notebook
-			.new_account_origins
-			.try_push(NewAccountOrigin::new(Alice.to_account_id(), AccountType::Tax, 4))
-			.expect("should insert");
-		assert_err!(
-			notebook_verify::<TestLookup>(&notebook.header.hash(), &notebook),
-			VerifyError::MissingBalanceProof
-		);
-		notebook.header.changed_accounts_root = merkle_root::<Blake2Hasher, _>(
-			balance_tips.iter().map(|(_, v)| v.encode()).collect::<Vec<_>>(),
-		);
-		notebook.balance_changes[1][0].previous_balance_proof = Some(BalanceProof {
-			notary_id: 1,
-			notebook_number: 1,
-			notebook_proof: None,
-			account_origin: AccountOrigin { notebook_number: 1, account_uid: 2 },
-			balance: 800,
-		});
-		notebook.balance_changes[1][1].previous_balance_proof = Some(BalanceProof {
-			notary_id: 1,
-			notebook_number: 1,
-			notebook_proof: None,
-			account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-			balance: 0,
-		});
-		notebook.balance_changes[1][2].previous_balance_proof = Some(BalanceProof {
-			notary_id: 1,
-			notebook_number: 1,
-			notebook_proof: None,
-			account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-			balance: 0,
-		});
-		assert_err!(
-			notebook_verify::<TestLookup>(&notebook.header.hash(), &notebook),
-			VerifyError::InvalidPreviousBalanceProof
-		);
-		notebook
-			.header
-			.changed_account_origins
-			.try_push(AccountOrigin { notebook_number: 1, account_uid: 4 })
-			.expect("should insert");
-
-		notebook.balance_changes[1][2].previous_balance_proof = None;
-		assert_ok!(notebook_verify::<TestLookup>(&notebook.header.hash(), &notebook),);
-	}
-
-	#[test]
-	fn test_cannot_remove_lock_between_changesets_in_a_notebook() {
-		let alice_balance_changeset = vec![BalanceChange {
-			balance: 1000,
-			change_number: 1,
-			account_id: Alice.to_account_id(),
-			account_type: AccountType::Deposit,
-			previous_balance_proof: None,
-			channel_hold_note: None,
-			notes: bounded_vec![Note::create(
-				1000,
-				NoteType::ClaimFromMainchain { account_nonce: 1 }
-			),],
-			signature: empty_signature(),
-		}
-		.sign(Alice.pair())
-		.clone()];
-		let alice_balance_changeset2 = vec![BalanceChange {
-			balance: 1000,
-			change_number: 2,
-			account_id: Alice.to_account_id(),
-			account_type: AccountType::Deposit,
-			previous_balance_proof: Some(BalanceProof {
-				notary_id: 1,
-				notebook_number: 1,
-				notebook_proof: None,
-				account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-				balance: 1000,
-			}),
-			channel_hold_note: None,
-			notes: bounded_vec![Note::create(
-				1000,
-				NoteType::ChannelHold { recipient: Bob.to_account_id() }
-			)],
-			signature: empty_signature(),
-		}
-		.sign(Alice.pair())
-		.clone()];
-
-		ValidLocalchainTransfers::mutate(|a| a.insert((Alice.to_account_id(), 1)));
-		let mut notebook = Notebook {
-			header: NotebookHeader {
-				version: 1,
-				notary_id: 1,
-				notebook_number: 1,
-				finalized_block_number: 100,
-				pinned_to_block_number: 0,
-				start_time: Utc::now().timestamp_millis() as u64 - 60_000,
-				changed_accounts_root: merkle_root::<Blake2Hasher, _>(vec![BalanceTip {
-					account_id: Alice.to_account_id(),
-					account_type: AccountType::Deposit,
-					balance: 1000,
-					change_number: 2,
-					account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-					channel_hold_note: None,
-				}
-				.encode()]),
-				chain_transfers: bounded_vec![ChainTransfer::ToLocalchain {
-					account_id: Alice.to_account_id(),
-					account_nonce: 1,
-				}],
-				changed_account_origins: bounded_vec![AccountOrigin {
-					notebook_number: 1,
-					account_uid: 1
-				}],
-				end_time: Utc::now().timestamp_millis() as u64,
-			},
-			balance_changes: bounded_vec![
-				BoundedVec::truncate_from(alice_balance_changeset),
-				BoundedVec::truncate_from(alice_balance_changeset2),
-			],
-			new_account_origins: bounded_vec![NewAccountOrigin::new(
-				Alice.to_account_id(),
-				AccountType::Deposit,
-				1
-			)],
-		};
-
-		// test that the change root records the hold note
-		assert_err!(
-			notebook_verify::<TestLookup>(&notebook.header.hash(), &notebook),
-			VerifyError::InvalidBalanceChangeRoot
-		);
-
-		let hold_note = notebook.balance_changes[1][0].notes[0].clone();
-
-		notebook.header.changed_accounts_root = merkle_root::<Blake2Hasher, _>(vec![BalanceTip {
-			account_id: Alice.to_account_id(),
-			account_type: AccountType::Deposit,
-			balance: 1000,
-			change_number: 2,
-			account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-			channel_hold_note: Some(hold_note),
-		}
-		.encode()]);
-		assert_ok!(notebook_verify::<TestLookup>(&notebook.header.hash(), &notebook),);
-
-		// now confirm we can't remove the hold in the same set of changes
-		{
-			// Try 1: pretend it didn't happen
-			let alice_balance_changeset3 = vec![BalanceChange {
-				balance: 1000,
-				change_number: 3,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: Some(BalanceProof {
-					notary_id: 1,
-					notebook_number: 1,
-					notebook_proof: None,
-					account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-					balance: 1000,
-				}),
-				channel_hold_note: None,
-				notes: bounded_vec![Note::create(
-					1000,
-					NoteType::ChannelHold { recipient: Ferdie.to_account_id() }
-				)],
-				signature: empty_signature(),
-			}
-			.sign(Alice.pair())
-			.clone()];
-			let mut notebook = notebook.clone();
-			let _ = notebook
-				.balance_changes
-				.try_push(BoundedVec::truncate_from(alice_balance_changeset3));
-			let hold_note = notebook.balance_changes[2][0].notes[0].clone();
-			notebook.header.changed_accounts_root =
-				merkle_root::<Blake2Hasher, _>(vec![BalanceTip {
-					account_id: Alice.to_account_id(),
-					account_type: AccountType::Deposit,
-					balance: 1000,
-					change_number: 3,
-					account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-					channel_hold_note: Some(hold_note),
-				}
-				.encode()]);
-			assert_err!(
-				notebook_verify::<TestLookup>(&notebook.header.hash(), &notebook),
-				VerifyError::InvalidChannelHoldNote
-			);
-		}
-		{
-			// Try 2: try to remove the hold
-			let alice_balance_changeset3 = vec![BalanceChange {
-				balance: 1000,
-				change_number: 3,
-				account_id: Alice.to_account_id(),
-				account_type: AccountType::Deposit,
-				previous_balance_proof: Some(BalanceProof {
-					notary_id: 1,
-					notebook_number: 1,
-					notebook_proof: None,
-					account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-					balance: 1000,
-				}),
-				channel_hold_note: Some(Note::create(
-					1000,
-					NoteType::ChannelHold { recipient: Bob.to_account_id() },
-				)),
-				notes: bounded_vec![Note::create(0, NoteType::ChannelSettle)],
-				signature: empty_signature(),
-			}
-			.sign(Alice.pair())
-			.clone()];
-
-			let mut notebook = notebook.clone();
-			let _ = notebook
-				.balance_changes
-				.try_push(BoundedVec::truncate_from(alice_balance_changeset3));
-			let hold_note = notebook.balance_changes[2][0].notes[0].clone();
-
-			notebook.header.changed_accounts_root =
-				merkle_root::<Blake2Hasher, _>(vec![BalanceTip {
-					account_id: Alice.to_account_id(),
-					account_type: AccountType::Deposit,
-					balance: 1000,
-					change_number: 3,
-					account_origin: AccountOrigin { notebook_number: 1, account_uid: 1 },
-					channel_hold_note: Some(hold_note),
-				}
-				.encode()]);
-			assert_err!(
-				notebook_verify::<TestLookup>(&notebook.header.hash(), &notebook),
-				VerifyError::ChannelHoldNotReadyForClaim
-			);
-		}
-	}
 }
